src/fiuba/algo3/algoformers/modelo/Escenario/Superficies/SuperficieTerrestre/TierraRocosa.java
package fiuba.algo3.algoformers.modelo.Escenario.Superficies.SuperficieTerrestre;

import fiuba.algo3.algoformers.modelo.Errores.NoPuedeInteractuarConSuperficieException;
import fiuba.algo3.algoformers.modelo.Escenario.Superficies.SuperficieTerrestre.SuperficieTerrestre;
import fiuba.algo3.algoformers.modelo.Personajes.AlgoFormer;
import fiuba.algo3.algoformers.modelo.Personajes.TiposDeUnidades.TipoUnidadAerea;

/**
 * Created by german.shokida on 27/5/2016.
 */
public class TierraRocosa implements SuperficieTerrestre {
    public void interactuar(AlgoFormer algoformer) throws NoPuedeInteractuarConSuperficieException {
        if(algoformer.esTipoUnidad(TipoUnidadAerea.getInstance()))
            throw new NoPuedeInteractuarConSuperficieException();
    }
}
